+++ cypress/integration/download.cy.js
@@ -1,14 +1,9 @@
-import {
-    DIMENSION_ID_ENROLLMENT_DATE,
-    DIMENSION_ID_EVENT_DATE,
-} from '../../src/modules/dimensionConstants.js'
-import { E2E_PROGRAM, TEST_REL_PE_LAST_YEAR } from '../data/index.js'
+import { E2E_PROGRAM } from '../data/index.js'
 import {
     selectEnrollmentProgram,
     selectEventWithProgram,
 } from '../helpers/dimensions.js'
 import { clickMenubarUpdateButton } from '../helpers/menubar.js'
-import { selectRelativePeriod } from '../helpers/period.js'
 import { goToStartPage } from '../helpers/startScreen.js'
 import { EXTENDED_TIMEOUT } from '../support/util.js'
 
@@ -37,15 +32,6 @@ describe('download', () => {
 
         clickMenubarUpdateButton()
 
-        downloadIsDisabled()
-
-        selectRelativePeriod({
-            label: E2E_PROGRAM[DIMENSION_ID_EVENT_DATE],
-            period: TEST_REL_PE_LAST_YEAR,
-        })
-
-        clickMenubarUpdateButton()
-
         downloadIsEnabled()
     })
 
@@ -56,15 +42,6 @@ describe('download', () => {
 
         selectEnrollmentProgram({
             programName: E2E_PROGRAM.programName,
-        })
-
-        clickMenubarUpdateButton()
-
-        downloadIsDisabled()
-
-        selectRelativePeriod({
-            label: E2E_PROGRAM[DIMENSION_ID_ENROLLMENT_DATE],
-            period: TEST_REL_PE_LAST_YEAR,
         })
 
         clickMenubarUpdateButton()