+++ cypress/support/index.js
@@ -1,6 +1,0 @@
-import { enableAutoLogin } from '@dhis2/cypress-commands'
-
-import './commands.js'
-
-enableAutoLogin()
-// enableNetworkShim()